Nice display of Table arguments steps
I'd like to change the way table arguments are display in order to make
it looks more like gherkin's syntaxe.

To to that, i'd suggest to use the StepParser to stringfy the tableargs
in json and transform it to a :
|     ArgA |     ArgB |
| Value A | Value B |

It will be far more readable in the output.

Just need to change the fullDescription to
exports.featureStepStringify(arguments[0], arguments[1], args),

(And add the reference to the file in karma-jasmine-cucumber)